requirements_test_osx.txt
-r requirements.txt
coverage==3.7.1
django-debug-toolbar==1.3.2
django-dynamic-fixture==1.8.5
ipdb==0.8.1
ipython==4.0.0
django-nose==1.4.1
